+++ Gravitas.Platform.Web/ViewModel/Admin/NodeTraffic/NodeTrafficListVm.cs
@@ -6,11 +6,11 @@ namespace Gravitas.Platform.Web.ViewModel
 {
     public class NodeTrafficListVm
     {
-        public long NodeId { get; set; }
+        public int NodeId { get; set; }
         [DisplayName("Назва вузла")]
         public string NodeName { get; set; }
         [DisplayName("Середній час у хвилинах")]
-        public long AverageTime { get; set; }
+        public int AverageTime { get; set; }
         [DisplayName("Машин в черзі")]
         public int ElementsInQueue { get; set; }
