+++ Gravitas.Platform.Web/ViewModel/Admin/Role/RoleVm.cs
@@ -6,7 +6,7 @@ namespace Gravitas.Platform.Web.ViewModel
     public class RoleVm
     {
         [DisplayName("Ідентифікатор ролі")]
-        public long RoleId { get; set; }
+        public int RoleId { get; set; }
         [DisplayName("Назва ролі")]
         public string RoleName { get; set; }
