+++ Gravitas.Platform.Web/ViewModel/Helper/GetTareValue.cs
@@ -4,9 +4,9 @@ namespace Gravitas.Platform.Web.ViewModel.Helper
 {
     public class GetTareValue
     {
-        public long NodeId { get; set; }
+        public int NodeId { get; set; }
         public TicketPackingTareVm[] TareItems { get; set; }
         public int OpRoutineReturnState { get; set; }
-        public long ReturnRoutineStateId { get; set; }
+        public int ReturnRoutineStateId { get; set; }
     }
 }
